+++ pkg-manager/plugin-commands-installation/package.json
@@ -46,7 +46,6 @@
     "@types/yarnpkg__lockfile": "catalog:",
     "@types/zkochan__table": "catalog:",
     "delay": "catalog:",
-    "@types/is-windows": "catalog:",
     "jest-diff": "catalog:",
     "path-name": "catalog:",
     "proxyquire": "catalog:",
@@ -101,7 +100,6 @@
     "ci-info": "catalog:",
     "enquirer": "catalog:",
     "is-subdir": "catalog:",
-    "is-windows": "catalog:",
     "load-json-file": "catalog:",
     "mem": "catalog:",
     "p-filter": "catalog:",
