+++ worker/src/index.ts
@@ -3,6 +3,8 @@ import path from 'path'
 import os from 'os'
 import { WorkerPool } from '@rushstack/worker-pool/lib/WorkerPool'
 import { PnpmError } from '@pnpm/error'
+import { execSync } from 'child_process'
+import isWindows from 'is-windows'
 import { type PackageFilesIndex } from '@pnpm/store.cafs'
 import { type DependencyManifest } from '@pnpm/types'
 import {
@@ -206,6 +208,21 @@ export async function importPackage (
   })
 }
 
+// In Windows system exFAT drive, symlink will result in error.
+function currentDriveIsExFAT (dir: string): boolean {
+  const currentDrive = `${dir.split(':')[0]}:`
+  const output = execSync(`wmic logicaldisk where "DeviceID='${currentDrive}'" get FileSystem`).toString()
+  const lines = output.trim().split('\n')
+  const name = lines.length > 1 ? lines[1].trim() : ''
+  return name === 'exFAT'
+}
+
+function extendMessage (err: any) { // eslint-disable-line @typescript-eslint/no-explicit-any
+  if (isWindows() && currentDriveIsExFAT(process.cwd())) {
+    err.message += '\n\nThe current drive is exFAT, which does not support symlinks. This will cause installation to fail. You can set the node-linker to "hoisted" to avoid this issue.'
+  }
+}
+
 export async function symlinkAllModules (
   opts: Omit<SymlinkAllModulesMessage, 'type'>
 ): Promise<{ isBuilt: boolean, importMethod: string | undefined }> {
@@ -217,6 +234,7 @@ export async function symlinkAllModules (
     localWorker.once('message', ({ status, error, value }: any) => { // eslint-disable-line @typescript-eslint/no-explicit-any
       workerPool!.checkinWorker(localWorker)
       if (status === 'error') {
+        extendMessage(error)
         reject(new PnpmError(error.code ?? 'SYMLINK_FAILED', error.message as string))
         return
       }
